==> notebooks/10 ranking_initial.ipynb <==
# Ranking: initial
In this notebook, we create an initial ranking metric based on the five key components predetermined to contribute to strong profitability of merchants to BNPL (earning growth rate, customer and merchant fraud probability, customer retention rate).

 These elements are equally weighted as a rudimetary ranking method. As performance and profit metrics from merchants in partnership with BNPL are not available, it is difficult to evaluate the accuracy of the ranking. In future sprints where this data is available, we can improve the model to generate more robust reccommendations.

from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.sql.types import *
import pandas as pd
from pyspark.ml.feature import VectorAssembler, StringIndexer, OneHotEncoder
from pyspark.ml.regression import LinearRegression
from pyspark.ml import Pipeline
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.ml.tuning import ParamGridBuilder, CrossValidator
from pyspark.ml.pipeline import PipelineModel
from pyspark.ml.regression import RandomForestRegressor
import matplotlib.pyplot as plt
import seaborn as sns
from pyspark.sql.window import Window

# Create a spark session
spark = (
    SparkSession.builder.appName("Initial Ranking Model")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.driver.memory", "12g")
    .config("spark.driver.maxResultSize", "16G")
    .config("spark.executor.memory", "16G")
    .config("spark.sql.files.maxPartitionBytes", "64MB")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.network.timeout", "600s")
    .getOrCreate()
)

We aim to create a function that will create the initial ranking. We will use the following:
- The earning growth rate
- The average customer fraud probability
- The merchant fraud probability
- The customer retention rate

We want all of these values to be between 0 and 1. Then we will give them a weight and sum them up to get the final ranking.

### Revenue Growth Rate

future_revenue = spark.read.parquet('../data/curated/predicted_monthly_revenue.parquet')
transactions = spark.read.parquet('../data/curated/all_details/')
merchant_monthly = transactions.groupBy(F.date_format(F.col('order_datetime'), 'yyyy-MM'), 'merchant_abn').agg(F.sum('dollar_value'))
merchant_monthly = merchant_monthly.withColumn('month', F.month(F.col('date_format(order_datetime, yyyy-MM)')))

# Creating a month since first transaction column for each merchant
merchant_monthly = merchant_monthly.withColumn('first_transaction', F.min('date_format(order_datetime, yyyy-MM)').over(Window.partitionBy('merchant_abn')))
merchant_monthly = merchant_monthly.withColumn('month_since_first_transaction', F.months_between(F.col('date_format(order_datetime, yyyy-MM)'), F.col('first_transaction')))
# Get the max month since first transaction for each merchant
max_month = merchant_monthly.groupBy('merchant_abn').agg(F.max('month_since_first_transaction').alias('max_month'))
max_month.show()

merchant_monthly = merchant_monthly.join(max_month, 'merchant_abn')
# Filter out the merchants with max month since first transaction less than 12
merchant_monthly = merchant_monthly.filter(F.col('month_since_first_transaction') >= 12)
# Filter out the last 12 months of transactions for each merchant
merchant_monthly = merchant_monthly.filter(F.col('month_since_first_transaction') >= F.col('max_month') - 12)
merchant_monthly.show()

# Sum the monthly revenue for each merchant
merchant_yearly = merchant_monthly.groupBy('merchant_abn').agg(F.sum('sum(dollar_value)').alias('yearly_revenue'))
merchant_yearly.show()

future_revenue.show()

# Sum the predicted monthly revenue for each merchant
summed_future_revenue = future_revenue.groupBy('merchant_abn').agg(F.sum('prediction').alias('predicted_yearly_revenue'))
summed_future_revenue.show()

# Add the take rate to summed future revenue
revenue = merchant_yearly.join(summed_future_revenue, 'merchant_abn')
revenue.show()

revenue = revenue.distinct()
revenue = revenue.withColumn('growth_rate', (F.col('predicted_yearly_revenue') - F.col('yearly_revenue')) / F.col('yearly_revenue'))
revenue.show()

### Customer Fraud Probability

customer_fraud = spark.read.parquet('../data/curated/predicted_consumer_fraud/*')
customer_fraud.show()

transactions.join(customer_fraud, 'order_id').show()

# Aggregate the average consumer fraud for each merchant
merchant_agg_fraud = transactions.groupby('merchant_abn').agg(F.avg('consumer_fraud').alias('avg_consumer_fraud'))

merchant_agg_fraud.show()

merchant_fraud = spark.read.parquet('../data/curated/predicted_merchant_fraud/*')

merchant_fraud.show()

order_abn = transactions.select('order_id', 'merchant_abn').distinct()

merchant_fraud = merchant_fraud.join(order_abn, on='order_id', how='left')

# Drop the order_id column
merchant_fraud = merchant_fraud.drop('order_id')
merchant_fraud.show()

### Customer Retention Rate

customer_retention = spark.read.parquet('../data/curated/customer_retention/*')
customer_retention.show()

## Creating the initial ranking


merchant_fraud = merchant_fraud.distinct()

# count the number of rows for each dataset before joining
print('Number of rows in revenue:', revenue.count())
print('Number of rows in merchant fraud:', merchant_fraud.count())
print('Number of rows in customer retention:', customer_retention.count())
print('Number of rows in customer fraud:', merchant_agg_fraud.count())

We can see that the number of records is same for all the datasets (4422) except for revenue dataset and merchant fraud. This is due to the fact that merchants with less than 12 months of data are not included in the revenue dataset. We will check the number of merchants in merchant fraud dataset.

# Counting the number of distinct merchants in merchant_fraud
print('Number of distinct merchants in merchant fraud:', merchant_fraud.select('merchant_abn').distinct().count())

# Showing the non distinct merchants in merchant_fraud
merchant_fraud.groupBy('merchant_abn').count().filter(F.col('count') > 1).show()

# Checking the values of merchant fraud for these non distinct merchants, we take the decision to take the average of the values
merchant_fraud = merchant_fraud.groupBy('merchant_abn').agg(F.avg('merchant_fraud').alias('merchant_fraud_prediction'))
merchant_fraud.count()

Weights will be 0.25 for each of the four factors.

ranking = revenue.join(merchant_agg_fraud, on ='merchant_abn', how = 'inner')
ranking = ranking.join(merchant_fraud, on = 'merchant_abn', how = 'inner')
ranking = ranking.join(customer_retention, on = 'merchant_abn', how = 'inner')
ranking.count()

# Scaling growth rate score to be between 0 and 1
min_growth_rate = ranking.agg(F.min('growth_rate')).collect()[0][0]
max_growth_rate = ranking.agg(F.max('growth_rate')).collect()[0][0]
ranking = ranking.withColumn('growth_rate_score', (F.col('growth_rate') - min_growth_rate) / (max_growth_rate - min_growth_rate))

ranking.show()

# Printing max and min values of growth rate score, average consumer fraud, merchant fraud and returning customer proportion to check if they are between 0 and 1
print('Max growth rate score:', ranking.agg(F.max('growth_rate_score')).collect()[0][0])
print('Min growth rate score:', ranking.agg(F.min('growth_rate_score')).collect()[0][0])
print('Max avg consumer fraud:', ranking.agg(F.max('avg_consumer_fraud')).collect()[0][0])
print('Min avg consumer fraud:', ranking.agg(F.min('avg_consumer_fraud')).collect()[0][0])
print('Max merchant fraud:', ranking.agg(F.max('merchant_fraud_prediction')).collect()[0][0])
print('Min merchant fraud:', ranking.agg(F.min('merchant_fraud_prediction')).collect()[0][0])
print('Max returning customer proportion:', ranking.agg(F.max('returning_customer_proportion')).collect()[0][0])
print('Min returning customer proportion:', ranking.agg(F.min('returning_customer_proportion')).collect()[0][0])

The consumer fraud probability and merchant fraud probability is to be divided by 100 to get the values between 0 and 1.

ranking = ranking.withColumn('final_score', (F.col('growth_rate_score') + (1-(F.col('avg_consumer_fraud')/100)) + (1-(F.col('merchant_fraud_prediction')/100)) + F.col('returning_customer_proportion'))/4)
ranking.show()

ranking = ranking.select('merchant_abn', 'final_score')


ranking=ranking.distinct()

ranking.sort(F.col('final_score').desc()).show()

ranking.write.mode('overwrite').parquet('../data/curated/merchant_ranking/')
