==> notebooks/13 ranking_final.ipynb <==
# Ranking: final
In this notebook, we combine our rankings from notebooks 10-12 to display a recommendation of the merchants to onbard to BNPL.

# Imports
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.sql.types import *
from pyspark.sql.window import Window

# Create Spark session
spark = (
    SparkSession.builder.appName("Merchant Ranking Model")
    .config("spark.sql.repl.eagerEval.enabled", True)
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.driver.memory", "12g")
    .config("spark.driver.maxResultSize", "16G")
    .config("spark.executor.memory", "16G")
    .config("spark.sql.files.maxPartitionBytes", "64MB")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.network.timeout", "600s")
    .getOrCreate()
)


# Load data
future_revenue = spark.read.parquet('../data/curated/predicted_monthly_revenue.parquet')
transactions = spark.read.parquet('../data/curated/all_details/')
merchant_segment = spark.read.parquet("../data/curated/merchant_segment.parquet")
merchant_data = spark.read.parquet("../data/curated/part_1/*")

# Define the merchant segments of interest
segments = [
    "Entertainment & Media",
    "Office & Home Supplies",
    "Miscellaneous",
    "Beauty",
    "Technology"
]

# getting rid of records with null values
merchant_data = merchant_data.na.drop()
merchant_data.count()

# Calculate the monthly revenue for each merchant
merchant_monthly = transactions.groupBy(F.date_format(F.col('order_datetime'), 'yyyy-MM'), 'merchant_abn').agg(F.sum('dollar_value'))
merchant_monthly = merchant_monthly.withColumn('month', F.month(F.col('date_format(order_datetime, yyyy-MM)')))

# Calculate the month since first transaction for each merchant
merchant_monthly = merchant_monthly.withColumn(
    'first_transaction', F.min('date_format(order_datetime, yyyy-MM)').over(Window.partitionBy('merchant_abn'))
)
merchant_monthly = merchant_monthly.withColumn(
    'month_since_first_transaction', F.months_between(F.col('date_format(order_datetime, yyyy-MM)'), F.col('first_transaction'))
)

# Get max month since the first transaction for each merchant
max_month = merchant_monthly.groupBy('merchant_abn').agg(F.max('month_since_first_transaction').alias('max_month'))

# Filter merchants with at least 12 months of data
merchant_monthly = merchant_monthly.join(max_month, 'merchant_abn')
merchant_monthly = merchant_monthly.filter(F.col('month_since_first_transaction') >= 12)

# Filter out the last 12 months of transactions for each merchant
merchant_monthly = merchant_monthly.filter(F.col('month_since_first_transaction') >= F.col('max_month') - 12)

# Sum the yearly revenue for each merchant
merchant_yearly = merchant_monthly.groupBy('merchant_abn').agg(F.sum('sum(dollar_value)').alias('yearly_revenue'))

# Sum predicted future monthly revenue for each merchant
summed_future_revenue = future_revenue.groupBy('merchant_abn').agg(F.sum('prediction').alias('predicted_yearly_revenue'))

# Join actual and predicted revenue
revenue = merchant_yearly.join(summed_future_revenue, 'merchant_abn')
revenue = revenue.withColumn('growth_rate', (F.col('predicted_yearly_revenue') - F.col('yearly_revenue')) / F.col('yearly_revenue'))

new_revenue = merchant_data.join(revenue, on='merchant_abn', how = 'left')
# Calculate growth in dollars
new_revenue = new_revenue.withColumn('growth_in_dollars', (F.col('predicted_yearly_revenue') - F.col('yearly_revenue'))*F.col('take_rate')/100)

new_revenue.sort('merchant_abn').show()

# Get merchant abn for each order id
order_abn = transactions.select('order_id', 'merchant_abn', 'order_datetime').distinct()

# Load customer fraud and merchant fraud data
consumer_fraud = spark.read.parquet('../data/curated/predicted_consumer_fraud/*')
consumer_fraud = consumer_fraud.join(order_abn, on='order_id', how='left')
agg_consumer_fraud = consumer_fraud.groupBy('merchant_abn').agg(F.avg('consumer_fraud').alias('consumer_fraud_prediction'))

# Load merchant fraud data
merchant_fraud = spark.read.parquet('../data/curated/predicted_merchant_fraud/*')
merchant_fraud = merchant_fraud.join(order_abn, on='order_id', how='left')

# Load customer retention data
customer_retention = spark.read.parquet('../data/curated/customer_retention/*')

merchant_fraud.show()

agg_consumer_fraud.show()

customer_retention.show()

# Merge the datasets
ranking = merchant_fraud.join(new_revenue, on='merchant_abn', how='left')
ranking = ranking.join(agg_consumer_fraud, on ='merchant_abn', how = 'left')
ranking = ranking.join(customer_retention, on = 'merchant_abn', how = 'left')

# Scale the growth rate and growth in dollars
min_growth_rate = ranking.agg(F.min('growth_rate')).collect()[0][0]
max_growth_rate = ranking.agg(F.max('growth_rate')).collect()[0][0]
ranking = ranking.withColumn('growth_rate_score', (F.col('growth_rate') - min_growth_rate) / (max_growth_rate - min_growth_rate))

min_growth_in_dollars = ranking.agg(F.min('growth_in_dollars')).collect()[0][0]
max_growth_in_dollars = ranking.agg(F.max('growth_in_dollars')).collect()[0][0]
ranking = ranking.withColumn('growth_dollars_score', (F.col('growth_in_dollars') - min_growth_in_dollars) / (max_growth_in_dollars - min_growth_in_dollars))

ranking.show()

# Create the final score by averaging all factors (weights can be adjusted if needed)
ranking = ranking.withColumn(
    'final_score', 
    (
        F.col('growth_rate_score') + 
        F.col('growth_dollars_score') + 
        (1 - (F.col('consumer_fraud_prediction') / 100)) + 
        (1 - (F.col('merchant_fraud') / 100)) + 
        F.col('returning_customer_proportion')
    ) / 5  # Averaging 5 factors: growth rate, growth dollars, fraud probabilities, and retention
)

# Join the merchant segments to the ranking
ranking_with_segment = ranking.join(merchant_segment, on='merchant_abn', how='inner')

ranking_with_segment.show()

# Select and sort the merchants by their final score
# We want to average the final score as well as merchant fraud
ranking_final = ranking_with_segment.groupby(
    'merchant_abn', 'name', 'revenue_level', 'consumer_fraud_prediction', 
    'returning_customer_proportion', 'growth_rate_score', 'growth_dollars_score', 'segment', 'growth_in_dollars'
).agg(F.avg('final_score').alias('average_final_score'), F.avg('merchant_fraud').alias('average_merchant_fraud'))
ranking_final = ranking_final.sort(F.col('average_final_score').desc())

ranking_final.show()

# Getting top 10 merchants for each segment
for segment in segments:
    print(f"Top 10 merchants for segment {segment}")
    ranking_final.filter(F.col('segment') == segment).show(10, truncate=False)
    print("\n\n")

# Checking how much revenue we gain by the top 100 merchants 
top_100 = ranking_final.limit(100)
top_100_sum = top_100.agg(F.sum('growth_in_dollars').alias('total_growth_in_dollars'))
top_100.show(truncate=False)

top_100_sum.show()

# Save the ranking
top_100.write.mode("overwrite").parquet("../data/curated/merchant_ranking")
