==> notebooks/15 summary.ipynb <==
# Summary notebook

Overall, our project has 6 stages: 
1. [Data downloading, cleaning and merging](#1-data-preprocessing)
2. [Imputation of transaction fraud probability using a predictive model](#2-impute-missing-transaction-fraud-probability)
3. [Creating predictive monthly revenue model](#3-prediction-of-future-monthly-revenue)
4. [Calculating consumer retention](#4-calculation-of-consumer-retention)
5. [Identifing 5 market segments](#5-identifying-segments)
6. [Ranking merchants](#6-calculation-of-final-scores-and-ranking)

from pyspark.sql import SparkSession
from pyspark.sql import functions as F
import shapefile as shp
import pandas as pd
import numpy as np
import os

# Create a spark session (which will run spark jobs)
spark = (
    SparkSession.builder.appName("Summary Notebook")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.driver.memory", "9g") 
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .getOrCreate()
)
spark.sparkContext.setLogLevel("OFF")

## 1. Data Preprocessing

In this stage, we first downloaded dataset from Canvas and ABS website. Then, we checked for duplicates, missing values, errors in data values, and carried out simple outlier analysis. Finally, we merged them together.

To summarize, we found that:
- Merchant, Consumer and transaction dataset are quite clean, so we only preprocessed merchant `tags` column and modified column names for consistency.
- For ABS Census data, we converted the each categorical value of type of measure column to a column. Moreover, we created a new feature based on the provided measures - weekly personal disposable.

# Merchant after cleaning
merchant = spark.read.parquet("../data/curated/part_1/tbl_merchants.parquet")
merchant.limit(5)

# External dataset after cleaning
abs_df = spark.read.parquet("../data/curated/sa2_dataset/abs_medians.parquet")
abs_df.limit(5)

# Final merged data
transactions = spark.read.parquet('../data/curated/all_details/')
transactions.limit(5)

## 2. Impute missing transaction fraud probability

After we merged all dataset together, we found many missing values as not all transaction was provided a merchant/consumer fraud probability. We decided to impute them using a Random Forest Regression model.

What we did:
- Doing some correlation analysis beforehand
- Doing feature selection and engineering, e.g `tags` column from merchant dataset.
- Evaluating model
- Making prediction

# Predicted consumer fraud probability
consumer_fraud = spark.read.parquet('../data/curated/predicted_consumer_fraud')
consumer_fraud.limit(5)

# Predicted merchant fraud probability
merchant_fraud = spark.read.parquet('../data/curated/predicted_merchant_fraud')
merchant_fraud.limit(5)

## 3. Prediction of future monthly revenue

- For generating merchant revenue growth rate, revenues were predicted for the following year of the obtained data.
- Separate models were ran for different revenue levels.

future_revenue = spark.read.parquet('../data/curated/predicted_monthly_revenue.parquet')
future_revenue.limit(5)

## 4. Calculation of consumer retention

- We assumed consumers with more than 1 transaction with the same merchant were returning customers of the merchant.
- For each merchant, returning customer proportion is calculated by dividing its number of returning customers by its total number of customers.

consumer_retention = spark.read.parquet("../data/curated/customer_retention/")
consumer_retention.limit(5)

## 5. Identifying segments

Based on provided type of goods in merchant dataset, we decided on 5 market segments: Entertainment & Media, Technology, Beauty (or Fashion), Office & Home Supplies, Miscellaneous.

We employed two distinct methods to identify them:
1. Members scanned and categorized each merchant by type of goods into five segments.
2. We applied the K-means clustering algorithm.

# Result of manual assignment
merchant_segment = spark.read.parquet("../data/curated/merchant_segment.parquet")
merchant_segment.limit(5)

# Result of Kmean clustering
merchant_segment = spark.read.parquet("../data/curated/clean_merchant_segmented.parquet")
merchant_segment.limit(5)

## 6. Calculation of final scores and Ranking

In this final stage, we took average of 5 metrics to get the final scores and got the top 100 merchants
- Consumer fraud probability
- Merchant fraud probability
- Merchant revenue
- Growth in Gross Earnings
- Consumer retention

About the Growth in Gross Earnings, we calculated it using merchant revenue.

top_100 = spark.read.parquet("../data/curated/merchant_ranking")
top_100.limit(5)

# Limitations and Assumptions

1. External consumer data are generalised and aggregated over postal regions, which may not be accurate or granular enough.
2. Due to limited data timeframe (2021-2022), ranking results of this project may not apply well to future years, where there might be changes in participating merchants, and updates to exsisting merchant data.
3. We used a simple Random Forest Regression model for imputation of missing transaction fraud probability, which may not be accurate.
4. Revenue growth rate is generated based on existing and modelled future data, due to limited data, which may have resulted in compounding inaccuracies.
5. For simplicity, geographical data mapping assumed each postcode region mapped to exactly one SA2 region, which is not entirely correct.
6. Modelling assumed that data being trained on were informative of their target data (fraud proababilities, revenue), and that the data size was sufficiently large for effective training.
